requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
